// src/multiplayer/generated/PveRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AiLevel } from "./AiLevel";
import type { FleetShip } from "./FleetShip";
import type { TaskGroup } from "./TaskGroup";

export type PveRequest = { version: number, seed: number, mapId: string, weather: string, difficulty: AiLevel, ships: Array<FleetShip>, groups: Array<TaskGroup>, };